Add files via upload
ML problem: Identifying URL is relevant or not.

For any textual data, all the categorical and textual data needs to be converted into numerical form before feeding to the machine learning algorithm, and that can be done by Word Embedding. It is an ML technique to convert a word into vectors. Some popular ways of doing that are one-Hot encoding, Bag of words, TF-IDF, Word2vec, FastText, etc. After that, the data set is ready to be applied to the ML algorithm. 

Steps:
1.   Loading of json files into data frames
2.   Combining those dataframes into one
3.   Shuffling it to prepare an unbiased data set
4.   Feature extraction
5.   Text Cleaning
6.   Tokenization
7.   Making the bag of words via sparse matrix
8.   Splitting of data set
9.   Fitting the training data to a random forest classifier
10.  Predicting the test data
11.  Generating classification report